Mudando tipo de preço para DECIMAL

+++ app/models.py
@@ -6,5 +6,5 @@ class Livro(db.Model):
     autor= db.Column(db.String(50), nullable=False)
     categoria= db.Column(db.String(30), nullable=True)
     editora= db.Column(db.String(40), nullable=True)
-    preco= db.Column(db.Integer, nullable=False)
+    preco= db.Column(db.Numeric(10, 2), nullable=False)
     ano_publicacao= db.Column(db.String(4), nullable=True)
